learning-v-esipov/src/main/java/com/aimprosoft/yesipov/web/command/impl/ListCommand.java
package com.aimprosoft.yesipov.web.command.impl;

import com.aimprosoft.yesipov.web.Path;
import com.aimprosoft.yesipov.web.command.Command;
import org.apache.log4j.Logger;

import javax.servlet.ServletException;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

public class ListCommand implements Command {

    private static final Logger log = Logger.getLogger(ListCommand.class);

    @Override
    public String execute(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
        log.debug("Command starts");

        String forward = request.getParameter("name").equals("departments") ?
                Path.DEPARTMENTS_JSP :
                Path.ALL_EMPLOYEES_JSP;

        log.trace("Forward address --> " + forward);
        log.debug("Controller finished, now go to forward address --> " + forward);

        log.debug("Command finished");
        return forward;
    }
}
